+++ task/config_test.go
@@ -1,20 +1,17 @@
 package task
 
 import (
+	"fmt"
 	"os"
 	"testing"
 	"time"
 
 	"github.com/circleci/ex/config/secret"
-	"github.com/circleci/ex/testing/testcontext"
-	"github.com/google/go-cmp/cmp/cmpopts"
 	"gotest.tools/v3/assert"
 	"gotest.tools/v3/assert/cmp"
 )
 
-func Test_ReadFromStdin(t *testing.T) {
-	ctx := testcontext.Background()
-
+func Test_UnmarshalJSON(t *testing.T) {
 	goodConfig := `
 {
 	"cmd": [],
@@ -24,14 +21,12 @@ func Test_ReadFromStdin(t *testing.T) {
 	"runner_api_base_url": "https://example.com/api",
 	"allocation": "testallocation",
 	"ssh_advertise_addr": "192.168.1.1",
-	"max_run_time": 60000000000,
-	"token_checksum": "ada63e98fe50eccb55036d88eda4b2c3709f53c2b65bc0335797067e9a2a5d8b"
+	"max_run_time": 60000000000
 }`
 	tests := []struct {
 		name string
 
 		rawConfig string
-		timeout   time.Duration
 
 		wantConfig Config
 		wantError  string
@@ -54,40 +49,15 @@ func Test_ReadFromStdin(t *testing.T) {
 			rawConfig: `not a valid JSON string`,
 			wantError: "failed to unmarshal config",
 		},
-		{
-			name:      "invalid checksum",
-			rawConfig: `{"token": "tasktoken","token_checksum": "invalid"}`,
-			wantError: "invalid checksum on config token",
-		},
-		{
-			name:      "timeout",
-			rawConfig: goodConfig,
-			timeout:   1 * time.Nanosecond,
-			wantError: "timed out reading config from stdin",
-		},
 	}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			config := Config{}
+			err := config.UnmarshalJSON([]byte(tt.rawConfig))
 
-			r, w, _ := os.Pipe()
-
-			go func() {
-				time.Sleep(100 * time.Millisecond)
-				_, err := w.Write([]byte(tt.rawConfig))
-				assert.NilError(t, err)
-				assert.NilError(t, w.Close())
-			}()
-
-			if tt.timeout == 0 {
-				tt.timeout = 20 * time.Second
-			}
-
-			err := config.ReadFrom(ctx, r, tt.timeout)
 			if tt.wantError == "" {
 				assert.NilError(t, err)
-				assert.Check(t, cmp.DeepEqual(config, tt.wantConfig, cmpopts.IgnoreFields(Config{}, "TokenChecksum")))
 			} else {
 				assert.Check(t, cmp.ErrorContains(err, tt.wantError))
 			}
@@ -118,7 +88,7 @@ func Test_Agent(t *testing.T) {
 	}
 
 	expectedEnv := []string{
-		"PATH=$PATH:/path/to",
+		fmt.Sprintf("PATH=%s:/path/to", os.Getenv("PATH")),
 	}
 
 	expectedAgent := Agent{